@@ -1,22 +1,35 @@
 {
-  "name": "jbook",
+  "name": "online_editor",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.3.0",
+    "@babel/parser": "^7.19.6",
+    "@babel/traverse": "^7.19.6",
+    "@monaco-editor/react": "^4.4.6",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.5.2",
-    "@types/node": "^16.11.45",
-    "@types/react": "^18.0.15",
+    "@types/jscodeshift": "^0.11.5",
+    "@types/prettier": "^2.7.1",
+    "@types/react": "^18.0.21",
     "@types/react-dom": "^18.0.6",
-    "axios": "^0.27.2",
-    "esbuild-wasm": "0.8.27",
+    "@types/react-monaco-editor": "^0.16.0",
+    "@types/react-resizable": "^3.0.3",
+    "axios": "^1.1.3",
+    "buffer": "^6.0.3",
+    "bulmaswatch": "^0.8.1",
+    "esbuild-wasm": "^0.15.11",
+    "jscodeshift": "^0.14.0",
     "localforage": "^1.10.0",
+    "monaco-editor": "^0.34.1",
+    "monaco-jsx-highlighter": "^2.0.4",
+    "prettier": "^2.7.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-scripts": "^5.0.1",
-    "typescript": "^4.7.4",
+    "react-resizable": "^3.0.4",
+    "react-resizeable": "^0.2.6",
+    "react-scripts": "5.0.1",
+    "typescript": "^4.8.4",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
